Add files via upload
Contains the annotation of all cases of it by two annotators, together with the word ID that corresponds to the GECO corpus. We also provide information of whether to two annotators agree on a case (1) or disagree on a case (0) in the column entitled match. The final two columns correspond to the words occurring left and right of the pronoun.